@@ -1,19 +1,19 @@
 [package]
 name = "edge-copilot-helper"
-version = "0.1.0"
-edition = "2021"
+version = "0.1.3"
+edition = "2024"
 description = "Cross-platform utility to bypass Microsoft Edge Copilot region restrictions"
 license = "MIT"
 repository = "https://github.com/qiyuey/edge-copilot-helper"
 
 [dependencies]
-serde_json = "1.0"
-dirs = "5.0"
-anyhow = "1.0"
+serde_json = "1"
+dirs = "5"
+anyhow = "1"
 clap = { version = "4", features = ["derive"] }
 
 [target.'cfg(not(target_os = "macos"))'.dependencies]
-sysinfo = "0.33"
+sysinfo = "0"
 
 [target.'cfg(target_os = "macos")'.dependencies]
 objc2 = "0.6"
