@@ -7,7 +7,7 @@ use sysinfo::System;
 use crate::common::apply_fix;
 
 #[cfg(target_os = "windows")]
-const PROCESS_NAMES: &[&str] = &["msedge.exe", "msedge"];
+const PROCESS_NAMES: &[&str] = &["msedge.exe"];
 
 #[cfg(target_os = "linux")]
 const PROCESS_NAMES: &[&str] = &[
@@ -23,7 +23,8 @@ const PROCESS_NAMES: &[&str] = &["msedge"];
 
 pub fn run_polling_loop() -> Result<()> {
     println!("🐧/🪟 Polling Mode: Starting Loop...");
-    println!("   Monitoring process: {}", PROCESS_NAMES.join(", "));
+    let process_list = PROCESS_NAMES.iter().copied().collect::<Vec<_>>().join(", ");
+    println!("   Monitoring process: {process_list}");
 
     let mut sys = System::new();
     let mut was_running = false;
@@ -31,17 +32,16 @@ pub fn run_polling_loop() -> Result<()> {
     loop {
         sys.refresh_processes(sysinfo::ProcessesToUpdate::All, true);
 
-        // Check if Edge is running (multiple processes, any one means Edge is open)
-        let is_running = PROCESS_NAMES.iter().any(|name| {
-            sys.processes()
-                .values()
-                .any(|p| p.name().to_string_lossy() == *name)
+        // Check if any Edge process exists
+        let is_running = sys.processes().values().any(|process| {
+            let pname = process.name().to_string_lossy();
+            PROCESS_NAMES.iter().any(|&n| n == pname)
         });
-
+        
         if was_running && !is_running {
             println!("🛑 Edge exited. Applying fix...");
             if let Err(e) = apply_fix() {
-                eprintln!("❌ Failed to apply fix: {}", e);
+                eprintln!("❌ Failed to apply fix: {e}");
             }
         }
 
