@@ -16,20 +16,20 @@ pub fn install() -> Result<()> {
     // 1. Create directories
     println!("Creating directories...");
     fs::create_dir_all(&install_dir)
-        .with_context(|| format!("Failed to create install directory: {:?}", install_dir))?;
+        .with_context(|| format!("Failed to create install directory: {}", install_dir.display()))?;
     fs::create_dir_all(&log_dir)
-        .with_context(|| format!("Failed to create log directory: {:?}", log_dir))?;
+        .with_context(|| format!("Failed to create log directory: {}", log_dir.display()))?;
 
     // Ensure LaunchAgents directory exists
     if let Some(parent) = plist_path.parent() {
         fs::create_dir_all(parent)
-            .with_context(|| format!("Failed to create LaunchAgents directory: {:?}", parent))?;
+            .with_context(|| format!("Failed to create LaunchAgents directory: {}", parent.display()))?;
     }
 
     // 2. Copy binary
     println!("Installing binary...");
     fs::copy(&current_exe, &binary_path)
-        .with_context(|| format!("Failed to copy binary to {:?}", binary_path))?;
+        .with_context(|| format!("Failed to copy binary to {}", binary_path.display()))?;
 
     // Make executable (should already be, but ensure it)
     #[cfg(unix)]
@@ -42,10 +42,11 @@ pub fn install() -> Result<()> {
 
     // 3. Unload existing service if present
     println!("Checking for existing service...");
+    let uid = get_uid();
     let _ = Command::new("launchctl")
         .args([
             "bootout",
-            &format!("gui/{}", get_uid()),
+            &format!("gui/{uid}"),
             plist_path.to_str().unwrap_or(""),
         ])
         .output();
@@ -54,7 +55,7 @@ pub fn install() -> Result<()> {
     println!("Creating Launch Agent plist...");
     let plist_content = generate_plist(&binary_path, &log_dir);
     fs::write(&plist_path, plist_content)
-        .with_context(|| format!("Failed to write plist to {:?}", plist_path))?;
+        .with_context(|| format!("Failed to write plist to {}", plist_path.display()))?;
 
     // 5. Load service
     println!("Loading service...");
@@ -69,10 +70,10 @@ pub fn install() -> Result<()> {
 
     println!();
     println!("Service installed and loaded successfully!");
-    println!("  Binary: {:?}", binary_path);
-    println!("  Logs:   {:?}", log_dir);
+    println!("  Binary: {}", binary_path.display());
+    println!("  Logs:   {}", log_dir.display());
     println!();
-    println!("Monitor with: tail -f {:?}/service.log", log_dir);
+    println!("Monitor with: tail -f {}/service.log", log_dir.display());
 
     Ok(())
 }
@@ -86,32 +87,33 @@ pub fn uninstall() -> Result<()> {
 
     // 1. Unload service
     println!("Stopping service...");
+    let uid = get_uid();
     let _ = Command::new("launchctl")
         .args([
             "bootout",
-            &format!("gui/{}", get_uid()),
+            &format!("gui/{uid}"),
             plist_path.to_str().unwrap_or(""),
         ])
         .output();
 
     // 2. Remove plist
     if plist_path.exists() {
-        println!("Removing plist: {:?}", plist_path);
+        println!("Removing plist: {}", plist_path.display());
         fs::remove_file(&plist_path)
-            .with_context(|| format!("Failed to remove {:?}", plist_path))?;
+            .with_context(|| format!("Failed to remove {}", plist_path.display()))?;
     }
 
     // 3. Remove install directory
     if install_dir.exists() {
-        println!("Removing files: {:?}", install_dir);
+        println!("Removing files: {}", install_dir.display());
         fs::remove_dir_all(&install_dir)
-            .with_context(|| format!("Failed to remove {:?}", install_dir))?;
+            .with_context(|| format!("Failed to remove {}", install_dir.display()))?;
     }
 
     // 4. Remove logs
     if log_dir.exists() {
-        println!("Removing logs: {:?}", log_dir);
-        fs::remove_dir_all(&log_dir).with_context(|| format!("Failed to remove {:?}", log_dir))?;
+        println!("Removing logs: {}", log_dir.display());
+        fs::remove_dir_all(&log_dir).with_context(|| format!("Failed to remove {}", log_dir.display()))?;
     }
 
     println!();
